added an option to feed inkscape with extra options

+++ inkscapeslide/__init__.py
@@ -49,10 +49,10 @@ layer name. The opacity must be between 0 and 1. Example:
     background, mylayer * 0.5
     """
     parser = OptionParser(usage=usage)
-    parser.add_option("-i", "--imageexport",
-            action="store_true", dest="imageexport", default=False,
-            help="Use PNG files as export content")
-    parser.add_option("-k", "--keep", action="store_true", dest="keeppdf", default=False, help="keep temporary pdf files (you can use these files in LaTeX/beamer with multiinclude")
+    parser.add_option("-i", "--imageexport", action="store_true", dest="imageexport", default=False, help="Use PNG files as export content")
+    parser.add_option("-k", "--keep", action="store_true", dest="keep_pdf", default=False, help="keep temporary pdf files (you can use these files in LaTeX/beamer with multiinclude")
+    parser.add_option("-o", "--inkscape-options", action="store",
+        dest="inkscape_opts", default="", help="with this option you can specify extra options that will be used for the extraction of the SVG file from inkscape")
     (options, args) = parser.parse_args()
     try:
         FILENAME = args[0]
@@ -174,9 +174,8 @@ layer name. The opacity must be between 0 and 1. Example:
         f.write(lxml.etree.tostring(doc))
         f.close()
 
-        # Determine whether to export pdf's or images (e.g. inkscape -A versus 
-        # inkscape -e)
-        cmd = "inkscape -D -A %s %s" % (pdfslide, svgslide)
+        # Determine whether to export pdf's or images (e.g. inkscape -A versus inkscape -e)
+        cmd = "inkscape %s -A %s %s" % (options.inkscape_opts, pdfslide, svgslide)
         if options.imageexport:
             cmd = "inkscape -D -d 180 -e %s %s" % (pdfslide, svgslide)
 
@@ -256,9 +255,10 @@ layer name. The opacity must be between 0 and 1. Example:
                     "python package, to join PDFs."
 
     # Clean up
-    if joinedpdf:
+    if (joinedpdf and (False == options.keeppdf)):
         for pdfslide in pdfslides:
             os.unlink(pdfslide)
+
 
 if __name__ == '__main__':
   main()
